Split apart `PackageExtension.swift` into individual source files.  Although it's a small special-purpose library, there is no reason to keep it all in one source files.  Internal declarations will work instead of fileprivates.
